+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to Pulselib will be documented in this file.
 
+## \[1.7.17] - 2024-03-21
+
+- Added scan_resonator_frequency for fast digitizer/rf_source frequency sweep on Qblox.
+- Log warning when sequence run duration > 3 s.
+
 ## \[1.7.16] - 2024-03-13
 
 - Use AWG channel attenuation for Keysight QS sequences.